pyproject.toml
[project]
name = "redletters"
version = "0.1.0"
description = "Multi-reading Greek NT source reader with interpretive receipts"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Red Letters Project"}
]
keywords = ["greek", "new-testament", "translation", "linguistics"]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "click>=8.1.0",
    "rich>=13.7.0",
    "pyyaml>=6.0",
    "sse-starlette>=1.6.0",
    "keyring>=24.0.0",
    "httpx>=0.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
    "ruff>=0.1.0",
]

[project.scripts]
redletters = "redletters.__main__:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"  # Explicit: avoid future pytest-asyncio default changes
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
# Warnings policy: OUR code fails on warnings, upstream stays noisy but visible
filterwarnings = [
    # Promote our own warnings to errors (catch regressions immediately)
    "error::DeprecationWarning:redletters.*",
    "error::PendingDeprecationWarning:redletters.*",
    "error::FutureWarning:redletters.*",
    "error::UserWarning:redletters.*",
    "error::ResourceWarning:redletters.*",
    # Let upstream deprecations stay warnings (don't whack-a-mole pydantic/httpx)
    "default::DeprecationWarning:pydantic.*",
    "default::DeprecationWarning:httpx.*",
]
